// src/models/polyglot.ts
export interface PolyglotMove {
  fromRow: number
  fromCol: number
  toRow: number
  toCol: number
  promotion?: 'queen' | 'knight' | 'rook' | 'bishop'
}